+++ src/util/crc32c.c
@@ -71,7 +71,7 @@ static inline uint32_t _x86_crc32c(uint32_t initial, const void *data, size_t le
     uint32_t crc = initial;
     for (size_t i = 0; i < len - (len % data_block_len); i += data_block_len) {
         #ifdef ARCH_X86_64
-        crc = _mm_crc32_u64(crc, *((uint64_t*) ((char*) data + i)));
+        crc = (uint32_t) _mm_crc32_u64((uint64_t) crc, *((uint64_t*) ((char*) data + i)));
         #else
         crc = _mm_crc32_u32(crc, *((uint32_t*) ((uintptr_t) data + i)));
         #endif
